requirements.txt
yfinance
textblob
sqlalchemy
pandas
matplotlib
streamlit
mysql-connector-python
beautifulsoup4
requests
altair
newsapi-python
plotly
schedule
